+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- `v0.0.16` SelectInput component
+  - Rework querying mecanism to respect user query
+  - Keep the selected option cached and always show it 
+
 ## 1.8.21 - 2025-05-15
 ### Fixed
 - `v0.0.15` SelectInput component